[Check] public_manuals_controller

+++ app/controllers/public/categories_controller.rb
@@ -4,9 +4,15 @@ class Public::CategoriesController < ApplicationController
   def create
     @category = Category.new(category_params)
     @category.user_id = current_user.id
-    @category.save
     @categories = Category.where(user_id: current_user.id)
-    redirect_to categories_path
+
+    if @category.save
+      flash[:notice] = "登録しました"
+      redirect_to categories_path
+    else
+      flash.now[:alert] = "登録できませんでした"
+      render :index
+    end
 
   end
 
@@ -23,15 +29,20 @@ class Public::CategoriesController < ApplicationController
     if @memos.count == 0
       @list_title = "「#{@category.name}」 の該当なし"
     else
-      @list_title = "#{@category.name} の検索結果一覧 （全 #{@memos.count}件）"
+      @list_title = "#{@category.name} の検索結果一覧 （全 #{@category.memos.count}件）"
     end
     render 'public/memos/index'
   end
 
   def destroy
     category = Category.find(params[:id])
-    category.destroy
-    redirect_to categories_path
+    if category.user_id == current_user.id
+      category.destroy
+      flash[:notice] = "削除しました"
+      redirect_to categories_path
+    else
+      redirect_to categories_path
+    end
   end
 
   def category_params
